pre-process: print variable argument macros correctly
The dump_macros() function fails to correctly output the definition of
macros that have a variable argument list. For example, the following
macros:

    #define unlocks(...) annotate(unlock_func(__VA_ARGS__))
    #define apply(x,...) x(__VA_ARGS__)

are output like so:

    #define unlocks(__VA_ARGS__) annotate(unlock_func(__VA_ARGS__))
    #define apply(x,__VA_ARGS__) x(__VA_ARGS__)

Add the code necessary to print the ellipsis in the argument list to the
dump_macros() function and add the above macros to the 'dump-macros.c'
test file.

Signed-off-by: Ramsay Jones <ramsay@ramsayjones.plus.com>
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>