cgcc: add support to ignore argument(s) of options
cgcc only does a minimal processing and filtering of its
command line and most options are simply forwarded to sparse
and gcc.

However, if one of the ignored options takes an argument that
matches one of the non-ignored options, this argument will be
processed as an option with undesirable effect.

Allow options to specify the number of arguments they're taking
and avoid any processing or filtering of these arguments while
still forwarding them to sparse and gcc.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>