cgcc: teach about '-x c'
Currently, cgcc only checks input files if their names end with
'.c' or if given as stdout. Other files are explicitly ignored.
This generally corresponds to what is wanted but GCC allows
arbitrary input files if the option '-x <language>' is given.
Some projects use this mechanism, for example to use the C
pre-processor on non-C files.

This fails when cgcc is used as wrapper around sparse + GCC.

Fix this by teaching cgcc about the '-x c' option.

Reported-by: Antonio Ospite <ao2@ao2.it>
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>