db: don't save struct members for integers
The problem is that system calls pass in an unsigned long, right? But it's
actually not an unsigned long, it's a pointer.  So now we dereference the
unsigned long and have information about the struct members.  That's all
good.

But say we take that unsigned long, and we pass it to a function which only
takes an int.  In that case, there is no way to dereference it so we don't
need to pass the struct members.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>