db: sql debug mode was executing the SQL twice
When you turned on debugging it was calling debug_sql() as well as
sql_exec() which executes the SQL twice.  That's harmless when you're
reading from the on-disk DB, but it becomes a problem when you're inserting
duplicate data into the in-mem DB.  Also it was really confusing to see the
duplicate data.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>