fix implicit K&R argument types
In an old-style function definition, if not explicitly specified,
the type of an argument defaults to 'int'.

Sparse issues an error for such arguments and leaves the type
as 'incomplete'. This can then create a cascade of other warnings.

Fix this by effectively giving the type 'int' to such arguments.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>