don't allow newlines inside string literals
Sparse allows (but warns about) a bare newline (not preceded by
a backslash) inside a string. Since this is invalid C, it's
probable that a terminating '"' is missing just before the newline.
In this case, allowing the newline implies accepting the following
characters until the next '"' is found, which is most case creates
a lot of irrelevant warnings.

Change this by disallowing newlines inside strings, exactly like
already done for character constants.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>