show-parse: don't display null ident in show_typename()
Often show_typename() is used to display a type and the associated
identifier is irrelevant but is displayed nevertheless.

However, when the identifier is itself not present, it is still
displayed as '<noident>', which is just noise and can be confusing.

Fix this by displaying nothing for null identifiers in show_typename().

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>