remove unneeded BSD & Solaris compatibility layer
None of the BSDs need "compat-bsd.c" anymore. Same for
Solaris and "compat-solaris.c", even for Solaris 10.

The only problem was lacking C99's strtold() but it seems
that this was solved many years ago and they're all doing
quite fine with "compat-linux.c".

So, simply replace the content of these file by an include of
"compat-linux.c".

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>