slist: handle the fast path in __merge_stree()
The normal thing when we're merging strees is for almost
all the states to stay the same.  It's not totally obvious
that merge_sm_states() and add_possible() are basically no-ops
in this case but this patch makes that more clear.

I would have thought that this patch would be a slightly speed
up but my benchmark says it is a small slow down.  Oh well.  At
least the code is more obvious.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>