db: don't create new splits in split_positive_from_negative()
The idea here is that maybe positive and negative mean
different things.  But the code here was splitting zero out
when maybe zero fits more naturally with the positives.

Another way of saying that is if you have an estate where
ret = -12 and another where ret = 0-32 then the "ret" state
should be divided at < zero and >= zero.  Otherwise we're just
creating an exact duplicate of the 1-32 states.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>