more consistent type info in error messages
Some error messages are displayed with auxillary information
about the concerned type(s).

However, this type information is displayed in various way:
just the type, "[left/right] side has type ...", "got ...", ...

Make these more consistent and simpler by just displaying
types when the error message is unambigous about the fact
that the problem is a type problem (and/or make the message
unambiguous when possible).

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>