extra: futz around with get_other_name_sym()
The issue here is that I had something like *(foo->bar)->baz but it's
replacing it with *(foo->bar).baz.  The || should clearly be an &&.  The
other line is less clear but also seems harmless.  And then a third issue
which isn't addressed is that actually the parenthesis are needed but I
strip them out.  And a fourth issue that isn't addressed is that I think
if we modify "*param->foo = val;" that might not be getting stored
correctly in the DB.  (I'm not sure.  But it didn't seem like it was at
first glance).

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>