asm: use parse_asm_constraint() to verify constraints
In extended ASM statements, output constraints need to be
prefixed with "=" or "+" and input constraints must not. This
is checked in verify_{output,input}_constraint() where the
constraint string is analyzed to look after these two chars.

However, the needed information is now already available thanks
to parse_asm_constraint().

So, use the result of the parsing of the constraint strings to
avoid to analyze again these strings during their verification.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>