asm: use a specific struct for asm operands
Before commit 756731e9 ("use a specific struct for asm operands")
ASM operands where stored as a list of n times 3 expressions.
After this commit, the triplets where stored inside a single
expression of type EXPR_ASM_OPERAND.

However, while this improved the parsing and use of ASM operands
it needlessly reuse 'struct expression' for something that is not
an expression at all.

Fix this by really using a specific struct for ASM operands.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>