equiv: use DB information to set extra state in equiv assignments
The implications of this are pretty severe.  Say we have:

	a = b->foo;
	if (a)
		__smatch_implied(a);

In the original code, if we didn't have a state saved for b->foo then we
set the implied state to the whole range.  But as well as looking at the
saved states we could also look at the DB information to find the value of
b->foo.

The real absolute code to handle assignments does use the DB information
and, in fact, when we get to "if (a) " then we use the DB information as
well.  So then when we're parsing the "if (a)" we say that "this can
never be non-NULL so there is nothing to do.  But when we call
__smatch_implied() it says "We have a state so there is no need to query
the DB.  We know this is the whole range."  So now it thinks that a can
be NULL despite the if statement and despite the information in the DB.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>