param_set: handle mtag assignments
The problem here is that we want to say that "$0->bar = 0-u64max[$1]",
meaning that we assign parameter 1 to foo->bar.  The problem is that if
it's an mtag, we would only record the mtag and not the [$1].  This causes
a problem for smatch_assigned_expr.c which needs the parameter.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>