Makefile: default to LD = CC
Usually the compiler is used as linker. Assuming that if someone wants
to change the compiler the linker should be changed, too, simplify that
use case by using "$(CC)" as linker instead of the hard coded "gcc".
This also matches the behaviour of make when using the built-in rules
of GNU Make which include:

	LINK.o = $(CC) $(LDFLAGS) $(TARGET_ARCH)
	%: %.o
		$(LINK.o) $^ $(LOADLIBES) $(LDLIBS) -o $@

Signed-off-by: Uwe Kleine-König <uwe@kleine-koenig.org>
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>