lib.c: move predefines out of handle_arch_m64_finalize()
In handle_arch_m64_finalize(), some types (like size_t_ctype)
are set to support the flag '-m64' and native 64-bit archs.
But some predefines are also issued.

As a preparatory step to fix a bug related to a inconstency
between this function and init_target(), move the predefines
to predefined_macros().

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>
Tested-by: Uwe Kleine-König <uwe@kleine-koenig.org>