cgcc: teach cgcc about Hurd/GNU
cgcc fails if it doesn't know about the system/OS as
returned by `uname -s`. This creates a build failure for
Debian since Hurd is one of their non-official 'ports'
but unknown to cgcc.

So, teach cgcc about 'GNU' (the OS/system name returned
on Hurd) and add the few predefines used to identify it.

Reported-by: Uwe Kleine-König <uwe@kleine-koenig.org>
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>