extra/untracked_param: mark lost parameters
Smatch has "untracked" parameters, which are parameters which are difficult
to track.  But that can be something simple like:

	foo->dev = dev;

And probably foo->dev is used for information only, it's not changed.  But
then there are functions where are too hairy to parse so we don't store any
information and those parameters are totally lost.  I've introduced a new
way to track those so that I can mark them as unknown in smatch_extra.

Lost parameters are untracked as well, obviously.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>