type: handle pointer type correctly
Smatch did a horrible hack for backwards compatibility but it turns out
that it was only needed for comparisons.  Also if you subtract an int from
a pointer then surely you get a pointer type?  I did that last change where
I switched the OR to and AND right before I committed the change so it
hasn't been tested.  But it feels right.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>