implied: don't use incomplete states
What happens here is that we have an if statement like so:

	if (as->usbm) {

The smatch_implied.c code thinks it can figure out what as->usbm should
be but it takes too long so it just says "0,4096-ptr_max".  Then the
smatch_extra.c code say it is 4096-ptr_max.  But then the implied code
over writes it with its own incomplete 0,4096-ptr_max state.  This patch
says that if we don't know then don't bother saving it.

I wanted to say that it was only necessary for the gate_sm, but when I
tested that it didn't work so I didn't investigate but just went with this
version.  It's hard for me to say if using incomplete states is ever really
useful...

I also merged a debugging change which is not 100% related.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>