helper: fix expr_to_chunk_helper() for struct members
I feel like expr_to_vsl() and expr_to_chunk_helper() should work the
same...  I'd have to think why they are different functions.

Anyway, expr_to_vsl() is slightly buggy.  And instead of fixing it, I'm
just going to use the fact that I already know the vsl in this case to
side step the bug and be a speed up.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>