buf_comparison: re-write to be more flexible
I've sort of sat on this commit for a while hoping that I would get
around to cleaning it up, but I haven't cleaned it up and now I'm just
committing it as-is.

Say we allocate an array like this:

	foo = alloc(size);

That's probably the most common way to allocate an array and the original
code was set up to handle that.  But sometimes, especially for strings,
people do:

	str = alloc(len + 1);
	...
	str[len] = '\0';

"len" is the last element of the array.

So there are different ways to describe the size of an array.  With
this patch the we track five types of relationships between the size
variable and the size of the buffer: byte_count, elem_count, last element,
number of elements used, and the last element used.

For the used elements that's when we do:

	array[num_elems++] = foo;

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>