equiv: update "other" names as well
The issue is that say we have another name then we do an assignment, I
guess we want to update the other name as well?

Ok.  It's not clear to me why we have another name and that's not already
marked as equiv.  Perhaps because we don't save those fake assignments?  Or
maybe the order is wrong.  Or maybe the get_other_name() does complicated
things with function arguments and breaking stuff into chunks so it
encompasses more than just equiv.

I wanted to spend some time on this and investigate but I haven't done
that.  This fixes a real life bug and I have been running it for a while
and I don't want to get out of sync with the released code.  As they say,
perfect is the enemy of barely acceptable.  :(

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>