testing_index_after_use: silence some false positives
I'm soon going to change how hard max works which triggered some new
false positives in this code, but really it's wrong to begin with.

We should be saying "Do we think that the index is valid?" but instead it
was saying that if we don't have a hard max treat it as invalid.  And it
was more strict when we were in non-spammy mode when the opposite was
intended.

There are several reasons why the index could be valid when we don't have
a hard max.  For example, the type could be u8 and the array is 256
elements.  Or we get the index from the DB but we know it's a valid range.

Also these days we can use the buf_comparison_index_ok() function to say if
we can use comparisons between functions instead of using literals to say
if the index is fine.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>