db: don't pretend to dereference integers
I recently noticed that a bunch of "ret->foo" dereferences were being
marked as capped.  The problem is "ret" was an int so it can't be
dereferenced.

I'm not sure how that happened.  Potentially something to do with error
pointers?  When I looked through the DB it had already stopped doing the
buggy behavior.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>