kernel: fix type bug handle IS_ERR() and friends
The problem was that llong types don't merge well with pointers.

[ Long side note]

I also considered removing all this code entirely.  The problem is that
when you call IS_ERR(ptr) where "ptr" is a known valid pointer then the
true return is deleted entirely and the PARAM_LIMIT information is deleted
so now if we have:

	if (IS_ERR(ptr))
		return PTR_ERR(ptr);

It says that we're passing a valid pointer to PTR_ERR().  The return is
marked impossible, so that's okay.  So one solution would be to not print
errors in impossible code...  But I decided that to keep this custom
IS_ERR() code instead.

I feel like ideally, we should be able to delete all this code, but it's
always been a complicated issue to figure out how to handle impossible
situations.  So I'm just kicking down the road for now.

[ End of long side note ]

The other thing that I changed here is that the code was
estate_filter_range() and that function was a bad design choice.  We should
only be manipulating range lists.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>