cgcc: favor using 'gcc -dumpmachine' to determine specifics
`uname -m` returns information about the host machine but this
information is useless when cgcc is used with a non-native compiler
since it's information about the target machine that is needed.

So, first try to determine the target machine via `gcc -dumpmachine`
and default to `uname -m`.

Note: this should fix problems with Debian build when armhf builder
   is run on a arm64 environment.

Originally-by: Uwe Kleine-König <uwe@kleine-koenig.org>
Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>