cgcc: simpler handling of hard-float ARM
There is an ABI for ARM with hard floats and one for soft floats
(as well as one for a sort of mix between hard & soft).
For hard floats, the preprocessor symbol '__ARM_PCS_VFP' needs to
be defined. This is added as an additional check in the code
returning the 'specs' for ARM.

To facilitate some upcoming changes and code reuse here, create
a pseudo-arch 'arm+hf' using '-D__ARM_PCS_VFP=1' in addition to
the usual options for ARM.

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>