zero_to_err_ptr: don't warn on impossible paths
I considered several options to addressing impossible IS_ERR() paths, and
in the end I decided to fix the type issue instead of removing all the
special IS_ERR() code from check_kernel.c...

But I may as well commit this as well.  It's a different way to solve the
same issue.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>