address: improve handling of __smatch_implied(&foo->bar->baz);
If you took the address of something and we couldn't figure it out, then
smatch_math.c was just marking it as a valid pointer.  That's probably
correct most of the time, but not 100% so it caused some false positives
where smatch said that an address couldn't be NULL so what's this NULL
check about?

So now the logic is:
1) If we can't figure out the offset of the member then return unknown
2) If we know that the foo->bar is non-NULL then just add the offset of
   baz to it and we're good.
3) If the offset is non-NULL but we can't figure out if foo->bar is non-NULL
   then just say that &foo->bar->baz is a valid pointer.
4) But if foo->bar is unknown and ->baz is zero offset, then the result is
   unknown and possibly NULL.

Ideally we would always be able to figure out the offset.  Or at least
return true if it's greater than zero...  But that's for a later patch.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>