slist: a pointer is equal to itself in cmp_possible_sm()
When I look at the cmp_possible_sm() function, I feel like it is a load
of absolute nonsense.  But then when I try to change it what I find is that
stuff breaks.

I'm trying to change it little by little.  This change is obviously safe
and should be a tiny bit of a speedup.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>