option: use handle_switches() for -m flags
The function handle_switch_m(), parsing the -m flags,
consists in a series of strcmp(), each setting a specific
value in one of the internal flag variable.

This can now be simplified by using a table to specify these
variables & values in a compact form.

So, convert the current code to use, if possible,  a table &
handle_switches().

Signed-off-by: Luc Van Oostenryck <luc.vanoostenryck@gmail.com>