implied: bail earlier on parsing parameter limitation implications
The problem here, is that eventually if we the function is too complicated
then we're going to have to bail.  We normally bail at the 1 minute mark.
(I should make this not a magic number).  Anyway, we could cut back on the
parameter implications slightly earlier so hopefully we don't have to bail
on everything.

People don't necessarily expect parameter implications to work anyway so
they probably wont' notice.  It's more "human like" to mess that up...
Also I was frustrated by a function which had tons and tons of implications
depending on which failure path we took, but nobody checked the failures.
Then we bailed on the obvious implications and printed a false positive...

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>