math: don't allocate the range_list if not necessary
The reason for this patch is that parsing
drivers/net/ethernet/mellanox/mlx5/core/en_main.c takes 30GB of RAM on
my system.  The reason for this is because the MLX5E_PARAMS_MINIMUM_LOG_RQ_SIZE
and MLX5E_PARAMS_DEFAULT_LOG_RQ_SIZE macros call ilog2() over and over.

The ilog2() is a very complicated macro and does tons of math, but a lot of
the time and in this case every value is known at compile time.  So instead
of generating tons and tons of range_lists we could just use an sval on the
stack.  This would save a lot of memory.

Except, unfortunately my benchmarking says that it actually makes the
problem worse.  None the less, I believe in the theory that this will help
eventually.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>