math: move more stuff to get_rl_sval()
This change can make a big improvement in memory use for some files which
call ilog2() a lot.  The thing about that macro is that there are a ton
of intermediary steps, but the value is always known.

The big things here are that I move get_implied_value_internal() and
get_value() to use get_rl_sval() instead of get_rl_helper().  For
get_value() we really shouldn't ever generate an rl.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>