address: improve get_address() for struct members
Ugh...  When I look at this change now, it's mostly just shifting code
around (clean ups).  I already committed the important bit that
get_member_offset_from_deref() wasn't working for pointer derefs.  But
this commit has a similar fix which is we call unop = strip_expr(unop->unop);
on pointers so get_implied_rl() has a better chance of working.

I had also considered saying that if implied_not_equal() and offset != 0
then mark it as valid.  But in the end, I didn't see a need because
hopefully, we're calculating the offset correctly most of the time now.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>