slist: reverse the order of the strees
Before smatch_extra was always at the end, but now I have moved it to
the start.  There are two reasons for this.  First it's technically
correct.  Secondly, I'm planning to change smatch_implied so that if we
only parse some of the implications then we'll use what we have.  I want
to put smatch_extra so it's more likely to get parsed before the cut off.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>