flow: fix valid_ptr_max
The problem here is that valid_ptr_max was being set to (LONG_MAX & ~OxfffUL)
instead of ULONG_MAX.

There was a little bit of fall out from this to do with error pointers.
Most of that was in code that I haven't committed, yet.  One small change
is that in check_kernel.c I was saying that that if it wasn't an error
pointer, that meant it was a valid pointer.  This is not true, because it
could also be a NULL pointer.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>