db/delete_too_common_fn_ptr.sh: delete some common function pointers
In the kernel, one of the files which takes the longest time to parse is
kernel/trace/blktrace.c and that's because it's trying to handle
'(struct tracepoint_func)->func' which gets passed to over a thousand
different functions.  We're not going to be able to handle it so just
remove it from the DB.

This removes about 7 other functions as well.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>