buf_size: fix handling of end of struct arrays that have one element
It used to be that if an array had one element it marked as unknown.  But
it turns out that it is useful to know about those.  The problem is that
I broke the handling for structs which end in a one element array.  This
was caught by the validation checks.

It turns the fix is very easy and we can just re-order
get_array_size_bytes_rl().  It shouldn't affect anything else.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>