kernel_user_data: handle struct assignments
This used to work, but just looking at the git log, I would have thought
it had been broken for forever...  Odd.  Anyway, after we do a struct
assignment from skb->data it was setting all the members to user data
and then clearing them  when we did the fake call assignment.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>