kernel: ignore lockdep information
This could have been done in fixup_kernel.sh but I had imagined I would
use the time information instead of just sed.  Also there is a lot of
data so it really is worthwhile to chop it out early.

The problem here is that we were recording that lockdep stuff changed
in "unknown ways" for every call.  No one was using the data.  Removing
it takes my DB from 32GB to 26GB so it saves 6GB.  Before we would hit
certain limits like if there are more than 3000 states changed in a return
then we just ignore it.  Also fewer states means everything is processed
faster.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>