implied: fix a bug handling partial implications
Partial implications are probably a bit bogus.  Probably the simpler thing
to do is to return NULL, but even handling that correct is slightly tricky
so, whatever, this fixes the bug and it's harmless.

Say we are trying to figure out the implications of a condition and run
out of time or hit a ->skip_implications flag.  Then we need to signal
back to filter_stack() that if this is the gate_sm() then we should not
save it.  If it's not the gate_sm, then we can use it, because any
implication is better than nothing.  But if it's the gate_sm, then we
should let smatch_extra.c handle the state.

Signed-off-by: Dan Carpenter <dan.carpenter@oracle.com>